refactor: modify login
