+++ src/main/java/com/scriptopia/demo/dto/user/RefreshResponse.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class TokenResponse {
+public class RefreshResponse {
     private String accessToken;
-    private String refreshToken;
+    private Long expiresIn;
 }
